Created me1130742.java
This application is a flight trip planner and computes the minimum cost for scheduling a trip from city A to city B given the departure and arrival times, along with time required in change of flights mid-way. Dijkstra's algorithm was implemented for this task. It had two versions of heaps. For smaller number of nodes in heaps, the priorities of the nodes could be changed. Fast linked list allowed for deletions in between the queue in O(1) given the previous node.